js/xpconnect/tests/chrome/test_asyncIteration.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1331092
-->
<window title="Mozilla Bug 1331092"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1331092"
     target="_blank">Mozilla Bug 1331092</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[
try {
    eval(`
async function* f() {
}
`);
    ok(false, "should throw");
} catch (e) {
    ok(e instanceof SyntaxError, "Async Generator shouldn't be supported in chrome code");
}

try {
    eval(`
async function f() {
  for await (let x of []) {
  }
}
`);
    ok(false, "should throw");
} catch (e) {
    ok(e instanceof SyntaxError, "for-await-of shouldn't be supported in chrome code");
}
  ]]></script>
</window>
